backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
web3>=6.15.0
python-dotenv>=1.0.1
pydantic>=2.6.0
